requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==4.0.3
django-environ==0.8.1
djangorestframework==3.13.1
install==1.3.5
pycodestyle==2.8.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
